FIXED issues with actions of the AdminPanel and changed the components folder within the components to subcomponents
