+++ frontend/src/components/AdminPanel/subcomponents/TabNavigation.jsx
