+++ frontend/src/components/Layout/subcomponents/Backdrop.jsx
